Fix build: move testData folder under /src, allow any again on trackingApi.fetchData.
