+++ packages/sitecore-jss/src/testData/mockDictionaryQueryResponse.json
