+++ packages/sitecore-jss/src/testData/mockSiteRootQueryResponse.json
